[DDW-889] Implement blocks component - fixes and project reorganization
