+++ source/features/blocks/actions.ts
@@ -1,4 +1,4 @@
-import Action from '../../../utils/Action';
+import Action from '../../utils/Action';
 
 export class SearchActions {
   public searchBlockById: Action<{ id: string }> = new Action();
