+++ source/features/blocks/api.ts
@@ -3,8 +3,8 @@ import ApolloClient from 'apollo-client';
 import {
   GetBlocksQuery,
   GetBlocksQueryVariables,
-} from '../../../../generated/typings/graphql-schema';
-import { GraphQLRequest } from '../../../utils/graphql/GraphQLRequest';
+} from '../../typings/graphql-schema';
+import { GraphQLRequest } from '../../utils/graphql/GraphQLRequest';
 import getBlocksQuery from './graphql/getBlocks.graphql';
 
 export class SearchApi {
