+++ source/features/epochs/components/EpochInfo.tsx
@@ -2,8 +2,8 @@ import { observer } from 'mobx-react-lite';
 import moment from 'moment';
 import CircularProgress, {
   CircularProgressSize,
-} from '../../shared/circular-progress/components/CircularProgress';
-import DividerWithTitle from '../../shared/divider-with-title/components/DividerWithTitle';
+} from '../../widgets/circular-progress/components/CircularProgress';
+import DividerWithTitle from '../../widgets/divider-with-title/components/DividerWithTitle';
 import styles from './EpochInfo.scss';
 
 export interface IEpochInfoProps {
