+++ source/features/widgets/circular-progress/components/CircularProgress.scss
