+++ source/features/widgets/circular-progress/components/CircularProgress.tsx
