+++ source/features/widgets/divider-with-title/components/DividerWithTitle.tsx
