+++ source/features/widgets/footer/components/Footer.scss
