+++ source/features/widgets/footer/components/Footer.tsx
