+++ source/features/widgets/header/components/Header.scss
