+++ source/features/widgets/search/components/Search.scss
