+++ source/features/widgets/search/graphql/getBlocks.graphql
