+++ stories/footer.stories.tsx
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import Footer from '../source/features/shared/footer/components/Footer';
+import Footer from '../source/features/widgets/footer/components/Footer';
 import { ThemeDecorator } from './support/ThemeDecorator';
 
 storiesOf('Footer', module)
